[wip] merge sight/obs form
