+++ frontend/src/app/programs/observations/form/form.component.ts
@@ -10,12 +10,14 @@ import { AppConfig } from '../../../../conf/app.config';
 
 declare let $: any;
 
+// TODO: migrate to conf
+export const taxonListThreshold = 10
 export const obsFormMarkerIcon = L.icon({
-    iconUrl:
-      "//cdnjs.cloudflare.com/ajax/libs/leaflet/1.2.0/images/marker-icon.png",
-    iconSize: [25, 40],
-    iconAnchor: [12, 40]
-  })
+  iconUrl: "../../../../assets/pointer-blue.png",  // TODO: Asset path should be normalized, conf ?
+  iconAnchor: [16, 42]
+})
+export const myMarkerTitle = '<i class="fa fa-eye"></i> Partagez votre observation'
+
 
 @Component({
   selector: 'app-sight-form',
@@ -37,9 +39,8 @@ export class SightsFormComponent implements AfterViewInit {
   surveySpecies: any;
   taxonomyList: any;
   program: any;
-  program_id: any = 1;
+  program_id: any;
   formMap: any
-  taxonListThreshold = 10
 
   constructor(
     private http: HttpClient,
@@ -57,8 +58,19 @@ export class SightsFormComponent implements AfterViewInit {
         this.getSurveySpeciesItems(this.taxonomyList);
         console.log('TaxonomyList', this.taxonomyList);
 
-        // TODO: constraints to keep program perimeter inbound (chckbx?)
+        // TODO: extract programArea centroid
         const formMap = L.map('formMap').setView([45.04499482319101, 5.042724609375001], 13)
+        /*
+         const formMap = L.map('formMap').locate({setView: true, maxZoom: 13})
+        function onLocationFound(e) {
+          const radius = e.accuracy / 2;
+          L.marker(e.latlng).addTo(map)
+              .bindPopup("You are within " + radius + " meters from this point").openPopup();
+
+          L.circle(e.latlng, radius).addTo(map);
+        }
+        formMap.on('locationfound', onLocationFound);
+        */
 
         L.tileLayer('//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
                 attribution: 'OpenStreetMap'
@@ -75,12 +87,13 @@ export class SightsFormComponent implements AfterViewInit {
             };
           }
         }).addTo(formMap)  // .bindPopup("Observation");
-        formMap.fitBounds(programArea.getBounds())
+
+        const maxBounds = programArea.getBounds()
+        formMap.fitBounds(maxBounds)
+        // QUESTION: enforce program area maxBounds (optional ?)
+        // formMap.setMaxBounds(maxBounds)
 
         let myMarker = null;
-
-        const myMarkerTitle =
-          '<i class="fa fa-eye"></i> Partagez votre observation';
 
         formMap.on("click", function(e) {
           //var Coords = "Lat, Lon : " + e.latlng.lat.toFixed(3) + ", " + e.latlng.lng.toFixed(3);
@@ -88,16 +101,25 @@ export class SightsFormComponent implements AfterViewInit {
             type: "Point",
             coordinates: [e.latlng.lng, e.latlng.lat]
           });
-          this.coords = coords;
+          // this.coords = coords;
           console.log(coords);
+
           if (myMarker !== null) {
             formMap.removeLayer(myMarker);
           }
-          myMarker = L.marker(e.latlng, { icon: obsFormMarkerIcon }).addTo(formMap);
-          $("#feature-title").html(myMarkerTitle);
-          $("#feature-coords").html(coords);
-          // $("#feature-info").html(myMarkerContent);
-          $("#featureModal").modal("show");
+
+          // PROBLEM: if program area is a concave polygon: one can still put a marker in the cavities.
+          // POSSIBLE SOLUTION: See ray casting algorithm for inspiration at https://stackoverflow.com/questions/31790344/determine-if-a-point-reside-inside-a-leaflet-polygon
+          if (L.latLngBounds(maxBounds).contains([e.latlng.lat, e.latlng.lng])) {
+            myMarker = L.marker(e.latlng, { icon: obsFormMarkerIcon }).addTo(formMap);
+            $("#feature-title").html(myMarkerTitle);
+            $("#feature-coords").html(coords);
+            // $("#feature-info").html(myMarkerContent);
+            $("#featureModal").modal("show");
+          /* } else {
+              console.debug(maxBounds, [e.latlng.lat, e.latlng.lng])
+          */
+          }
         });
       })
   }
